Add infra node
Dedicated DNS node is replaced with infra node which is now used
for running openshift-ansible (so far openshift-ansible was running
on one of master nodes) and for running DNS.

Running openshift-ansible from "other than master" node is needed
for the upcoming containerization patch: openshift-ansible restarts
docker service during its run, if openshift-ansible runs in container
itself it's being restarted too.

Having openshift-ansible on infra node is also cleaner in HA mode
when it's not clear on which of master nodes openshift-ansible runs.